+++ tests/testthat/test_readWriteMgfData.R
@@ -19,12 +19,16 @@ test_that("extractMgfSpectrum2Info", {
            mz = c(10, 11, 12, 13),
            intensity = c(100, 200, 300, 400),
            fromFile = 1L,
+           acquisitionNum = NA_integer_,
+           collisionEnergy = NA_real_,
+           precScanNum = NA_integer_,
            centroided = TRUE)
-  
+
   fdata <- c(TITLE="foobar File=\"foobar.raw\", Native ID:\"controllerType=0 controllerNumber=1, scan=100\"",
              RTINSECONDS="600",
              PEPMASS="100 50000",
              CHARGE="3+")
   result <- list(spectrum=s, fdata=fdata)
-  expect_identical(MSnbase:::extractMgfSpectrum2Info(mgf, centroided = TRUE), result)
+  expect_equal(MSnbase:::extractMgfSpectrum2Info(mgf, centroided = TRUE),
+               result)
 })