+++ NEWS.md
@@ -1,8 +1,7 @@
 # MSnbase 2.5
 
 ## Changes in version 2.5.8
-- Setting default cv to FALSE in combineFeatures (see issue #303)
-  <2018-02-08 Thu>
+- Nothing yet
 
 ## Changes in version 2.5.7